Extract codec in it's own crate
