+++ tremor-cli/src/errors.rs
@@ -43,6 +43,7 @@ error_chain! {
         Script(tremor_script::errors::Error, tremor_script::errors::ErrorKind);
         Pipeline(tremor_pipeline::errors::Error, tremor_pipeline::errors::ErrorKind);
         Runtime(tremor_runtime::errors::Error, tremor_runtime::errors::ErrorKind);
+        Codec(tremor_codec::errors::Error, tremor_codec::errors::ErrorKind);
     }
     foreign_links {
         Value(tremor_value::Error);
