+++ tremor-codec/Cargo.toml
@@ -1,0 +1,54 @@
+[package]
+authors = ["The Tremor Team"]
+name = "tremor-codec"
+edition = "2021"
+description = "Tremor Script Interpreter"
+documentation = "https://docs.rs/tremor-script"
+license = "Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/tremor-rs/tremor-runtime"
+version = "0.13.0-rc.15"
+
+
+[dependencies]
+tremor-value = { version = "0.13.0-rc.15", path = "../tremor-value" }
+tremor-common = { version = "0.13.0-rc.15", path = "../tremor-common" }
+tokio = { version = "1.32", features = ["full"] }
+async-trait = "*"
+error-chain = "0.12"
+futures = "*"
+simd-json-derive = "*"
+value-trait = "*"
+beef = "*"
+test-case = "3.1"
+proptest = "1.1"
+
+#support
+byteorder = "*"
+itoa = "1"
+ryu = "1"
+lexical = "6"
+simdutf8 = "0.1"
+chrono = "0.4"
+chrono-tz = "0.8"
+uuid = { version = "1.3", features = ["v4"] }
+matches = "0.1"
+
+
+# codecs
+tremor-influx = { version = "0.13.0-rc.15", path = "../tremor-influx" }
+simd-json = "*"
+apache-avro = { version = "0.15", features = [
+    "snappy",
+    "bzip",
+    "xz",
+    "zstandard",
+    "bzip2",
+] }
+serde = "*"
+reqwest = "*"
+csv = "1.2"
+rmp-serde = "1.1"
+syslog_loose = "0.19"
+serde_yaml = "0.9"
+hdrhistogram = "7"
