+++ tremor-codec/src/codec/null.rs
@@ -12,46 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! The `msgpack` codec supports the [Msgpack](https://msgpack.org) binary format that.
-//!
-//! The format is structurally compatible with JSON.
-//!
-//! Message pack is typically significantly more compute efficient to process and requires less memory
-//! when compared to typical JSON processors.
+//! The `null` codec drops data.
 
-use super::prelude::*;
-use rmp_serde as rmps;
+use crate::prelude::*;
 
 #[derive(Clone)]
-pub struct MsgPack {}
+pub struct Null {}
 
 #[async_trait::async_trait]
-impl Codec for MsgPack {
+impl Codec for Null {
     fn name(&self) -> &str {
-        "msgpack"
-    }
-
-    fn mime_types(&self) -> Vec<&'static str> {
-        vec![
-            "application/msgpack",
-            "application/x-msgpack",
-            "application/vnd.msgpack",
-        ]
+        "null"
     }
 
     async fn decode<'input>(
         &mut self,
-        data: &'input mut [u8],
+        _data: &'input mut [u8],
         _ingest_ns: u64,
         meta: Value<'input>,
     ) -> Result<Option<(Value<'input>, Value<'input>)>> {
-        rmps::from_slice::<Value>(data)
-            .map(|v| Some((v, meta)))
-            .map_err(Error::from)
+        Ok(Some((Value::const_null(), meta)))
     }
-
-    async fn encode(&mut self, data: &Value, _meta: &Value) -> Result<Vec<u8>> {
-        Ok(rmps::to_vec(&data)?)
+    async fn encode(&mut self, _data: &Value, _meta: &Value) -> Result<Vec<u8>> {
+        Ok(vec![])
     }
 
     fn boxed_clone(&self) -> Box<dyn Codec> {
@@ -62,18 +45,20 @@ impl Codec for MsgPack {
 #[cfg(test)]
 mod test {
     use super::*;
-    use tremor_value::literal;
+    use simd_json::OwnedValue;
 
     #[tokio::test(flavor = "multi_thread")]
-    async fn test_msgpack_codec() -> Result<()> {
-        let seed = literal!({ "snot": "badger" });
+    async fn test_null_codec() -> Result<()> {
+        let seed: OwnedValue = OwnedValue::null();
+        let seed: Value = seed.into();
 
-        let mut codec = MsgPack {};
+        let mut codec = Null {};
         let mut as_raw = codec.encode(&seed, &Value::const_null()).await?;
-        assert!(codec
+        let as_json = codec
             .decode(as_raw.as_mut_slice(), 0, Value::object())
-            .await?
-            .is_some());
+            .await;
+        assert!(as_json.is_ok());
+        as_json?;
 
         Ok(())
     }
